fix types in disarium_number.cu

+++ CUDA/kernels/disarium_number.cu
@@ -1,8 +1,8 @@
-__device__ unsigned long countDigits(unsigned long number);
+__device__ unsigned int countDigits(unsigned long number);
 
 __device__ bool isNumberDisarium(unsigned long number);
 
-__device__ unsigned long pow(unsigned long x, unsigned long n);
+__device__ unsigned long pow(unsigned long x, unsigned int n);
 
 __device__ void addResult(unsigned long *generatedNumbersGPU, unsigned long result);
 
@@ -23,7 +23,7 @@ __device__ void addResult(unsigned long *generatedNumbersGPU, unsigned long resu
 
 __device__ bool isNumberDisarium(unsigned long number) {
     unsigned long sum = 0, temp = number;
-    unsigned long digitsCount = countDigits(number);
+    unsigned int digitsCount = countDigits(number);
     while (temp) {
         sum += pow(temp % 10, digitsCount--);
         temp /= 10;
@@ -31,7 +31,7 @@ __device__ bool isNumberDisarium(unsigned long number) {
     return sum == number;
 }
 
-__device__ unsigned long countDigits(unsigned long number) {
+__device__ unsigned int countDigits(unsigned long number) {
     unsigned int digitsCount = 0;
     while (number) {
         number /= 10;
@@ -40,7 +40,7 @@ __device__ unsigned long countDigits(unsigned long number) {
     return digitsCount;
 }
 
-__device__ unsigned long pow(unsigned long x, unsigned long n) {
+__device__ unsigned long pow(unsigned long x, unsigned int n) {
     unsigned long result = 1;
     for (unsigned int i = 0; i < n; i++)
         result *= x;
